Merged 'companies with' and 'companies without' job sections.
This allows for easier community maintenance (less moving companies between sections).

Additionally, I did add a wording change to the 'co-op friendly' section that may not be appropriate - stating that these companies are open to interns as well. It was added because I noticed intern positions being listed in the 'companies with jobs' section, which dilutes the nature of the list, in my opinion. If not appropriate, I can remove, no worries.